requirements.txt
icrawler
duckduckgo-search
requests
pillow
torch
transformers
tqdm
pandas 